requirements.txt
torch==1.10.0
aiohttp==3.8.1
einops==0.3.2
ftfy==6.0.3
matplotlib==3.5.1
mpi4py==3.1.3
numpy==1.21.5
omegaconf==2.1.1
filelock==3.4.2
packaging==21.3
Pillow==8.4.0
pytorch-lightning==1.5.7
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
termcolor==1.1.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
wandb==0.12.9